Move to a top level conftest.py
